# api/swagger.yaml
basePath: /api/v1
definitions:
  internal_api_handler.AggregateRequest:
    properties:
      field:
        type: string
    required:
    - field
    type: object
  internal_api_handler.AggregateResponse:
    properties:
      buckets:
        additionalProperties:
          format: int64
          type: integer
        type: object
      field:
        type: string
      success:
        type: boolean
    type: object
  internal_api_handler.CountResponse:
    properties:
      count:
        type: integer
      success:
        type: boolean
    type: object
...
  internal_api_handler.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  internal_api_handler.HybridSearchRequest:
    properties:
      filters:
        additionalProperties: true
        type: object
      k:
        type: integer
      query:
        type: string
      vector:
        items:
          type: number
        type: array
      vector_field:
        type: string
    required:
    - query
    type: object
  internal_api_handler.KNNSearchRequest:
    properties:
      field:
        description: 向量字段名：content_vector, image_vector
        type: string
      filters:
        additionalProperties: true
        description: 过滤条件
        type: object
      k:
        description: 返回结果数量
        type: integer
      vector:
        items:
          type: number
        type: array
    required:
    - vector
    type: object
  internal_api_handler.KNNSearchResponse:
    properties:
      hits:
        items:
          $ref: '#/definitions/internal_api_handler.VectorHit'
        type: array
      success:
        type: boolean
      took_ms:
        type: integer
      total:
        type: integer
    type: object
...
  internal_api_handler.SearchHit:
    properties:
      highlight:
        additionalProperties: true
        type: object
      id:
        type: string
      score:
        type: number
      source:
        additionalProperties: true
        type: object
    type: object
  internal_api_handler.SearchRequest:
    properties:
      filters:
        additionalProperties: true
        type: object
      from:
        type: integer
      highlight:
        additionalProperties: true
        type: object
      query:
        type: string
      size:
        type: integer
      sort:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
  internal_api_handler.SearchResponse:
    properties:
      hits:
        items:
          $ref: '#/definitions/internal_api_handler.SearchHit'
        type: array
      success:
        type: boolean
      took_ms:
        type: integer
      total:
        type: integer
    type: object
...
  internal_api_handler.VectorHit:
    properties:
      id:
        type: string
      score:
        type: number
      source:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    多租户文件索引服务，支持 PDF、图片、文档等文件的索引、存储和搜索
    认证使用 JWT Token，数据按租户隔离存储在独立的 OpenSearch 索引中。
  title: OpenSearch File API
  version: 1.0.0
paths:
...
  /search:
    get:
      description: 使用 URL 查询参数进行搜索
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 文件类型
        in: query
        name: file_type
        type: string
      - description: MIME 类型
        in: query
        name: content_type
        type: string
      - default: 0
        description: 起始位置
        in: query
        name: from
        type: integer
      - default: 10
        description: 返回数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 搜索文件（GET）
      tags:
      - Search
    post:
      consumes:
      - application/json
      description: 使用 JSON 请求体进行全文搜索、过滤和排序
      parameters:
      - description: 搜索请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 搜索文件（POST 高级搜索）
      tags:
      - Search
  /search/aggregate:
    post:
      consumes:
      - application/json
      description: 按指定字段聚合统计
      parameters:
      - description: 聚合请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.AggregateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.AggregateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 聚合查询
      tags:
      - Search
  /search/count:
    get:
      description: 统计当前租户下的文件总数
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.CountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 统计文件数量
      tags:
      - Search
  /search/hybrid:
    post:
      consumes:
      - application/json
      description: 结合文本关键词和向量进行混合搜索，同时匹配文本相关性和向量相似性
      parameters:
      - description: 混合搜索请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.HybridSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.KNNSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: 混合搜索（文本 + 向量）
      tags:
      - Vector Search
  /search/knn:
    post:
      consumes:
      - application/json
      description: 使用向量进行 K 近邻搜索，支持 content_vector (1536维) 和 image_vector (512维)
      parameters:
      - description: KNN 搜索请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.KNNSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.KNNSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      - X-Tenant-ID: []
      summary: KNN 向量搜索
      tags:
      - Vector Search
...
securityDefinitions:
  BearerAuth:
    description: JWT Token，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  X-Tenant-ID:
    description: 租户 ID
    in: header
    name: X-Tenant-ID
    type: apiKey
swagger: "2.0"